% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{activeBreakpoints}
\alias{activeBreakpoints}
\title{Current active media breakpoints. By default the default_system
  setting values are used, but the list can be customized by using the
  registerBreakpoint() and unregisterBreakpoint() functions.}
\usage{
activeBreakpoints()
}
\value{
A named list of media breakpoints options.
}
\description{
Current active media breakpoints. By default the default_system
  setting values are used, but the list can be customized by using the
  registerBreakpoint() and unregisterBreakpoint() functions.
}
