% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{addBreakpoint}
\alias{addBreakpoint}
\title{Adds a breakpoint to a breakpoint system object.}
\usage{
addBreakpoint(system, breakpoint)
}
\arguments{
\item{system}{A breakpoint system object generated with breakpointSystem().}

\item{breakpoint}{A valid breakpoint generated with breakpoint().}
}
\value{
A breakpoint system object.
}
\description{
Adds a breakpoint to a breakpoint system object.
}
\keyword{breakpoints}
