% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addClass}
\alias{addClass}
\title{Add object class}
\usage{
addClass(object, class)
}
\arguments{
\item{object}{Any R object.}

\item{class}{A string representing a object class.}
}
\value{
The given R object with the additional class.
}
\description{
Adds a class to a object
}
\keyword{internal}
\keyword{utils}
