% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{applyTemplate}
\alias{applyTemplate}
\title{Apply a template to a attribute list}
\usage{
applyTemplate(attributes, template, defaults, type)
}
\arguments{
\item{attributes}{The manually given attribute values that will take priority
during the merge.}

\item{template}{The name of the template to merge, or a template object.}

\item{defaults}{Default attribute values to ignore from attributes.}

\item{type}{The type of css grid of the template.}
}
\value{
A named list of css attributes that can be used to generate a html
element style rules of the given type.
}
\description{
Merges a set of attributes with a given template. To avoid redundant
attributes being added to the final list, a list of default values (based
of the specific panel creation callback formals) is used to validate the
need of a argument value in the final list.
}
\keyword{templates}
