% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{breakpoint}
\alias{breakpoint}
\title{Generated a valid breakpoint to use in a breakpoint system. While both the
min and max arguments are optional, at least one of them must exist for
the breakpoint to be considered valid.}
\usage{
breakpoint(name, min = NULL, max = NULL)
}
\arguments{
\item{name}{The name of the breakpoint.}

\item{min}{Optional numeric minimum value (in pixels) of the screen width
where the breakpoint is valid.}

\item{max}{Optional numeric maximum value (in pixels) of the screen width
where the breakpoint is valid.}
}
\value{
A breakpoint object.
}
\description{
Generated a valid breakpoint to use in a breakpoint system. While both the
min and max arguments are optional, at least one of them must exist for
the breakpoint to be considered valid.
}
\keyword{breakpoints}
