% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{breakpointSystem}
\alias{breakpointSystem}
\title{Generated a valid breakpoint system to use in grid functions.}
\usage{
breakpointSystem(name, ..., description = NULL)
}
\arguments{
\item{name}{The name of the breakpoint system.}

\item{...}{Any number of breakpoint objects generated with breakpoint().}

\item{description}{Optional description with additional information
regarding the breakpoint system.}
}
\value{
A breakpoint system object.
}
\description{
Generated a valid breakpoint system to use in grid functions.
}
\keyword{breakpoints}
