% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{exportBreakpointSystem}
\alias{exportBreakpointSystem}
\title{Exports a breakpoint system into a file for storage and later usage.
Exported files can be retrieved from thier file form by using
importBreakpointSystem()}
\usage{
exportBreakpointSystem(system, path)
}
\arguments{
\item{system}{The name of a registered breakpoint system, or a breakpoint
system object generated with breakpointSystem().}

\item{path}{A file path where to export the system to including the file
name and extension. The file name must end with a ".yaml" extension.}
}
\value{
No return value, called for side effects.
}
\description{
Exports a breakpoint system into a file for storage and later usage.
Exported files can be retrieved from thier file form by using
importBreakpointSystem()
}
\keyword{breakpoints}
