% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{flexPage}
\alias{flexPage}
\title{Create a css flexbox based page}
\usage{
flexPage(..., title = NULL, fill_page = TRUE, dependency = bootstrapLib())
}
\arguments{
\item{...}{Arguments to be passed to \link{flexPanel}.}

\item{title}{The browser window title (defaults to the host URL of the page).}

\item{fill_page}{Boolean value if the page should automatically stretch to
match the browser window height.}

\item{dependency}{A set of web dependencies. This value can be a
\link[htmltools:htmlDependency]{htmltools::htmlDependency}, for example the shiny bootstrap dependency (default
value) or a \link[htmltools:tagList]{htmltools::tagList} with different dependencies.}
}
\value{
A UI definition that can be passed to the \link[shiny:shinyUI]{shiny::shinyUI} function.
}
\description{
Create a Shiny UI page using a \link{flexPanel} to wrap the page content.
As other Shiny UI pages, it scaffolds the entire page and loads any required
or registered dependencies.
}
\note{
See
\url{https://css-tricks.com/snippets/css/a-guide-to-flexbox/}
for additional details on using css flexbox.
}
\seealso{
\link{flexPanel}
}
\concept{flex}
\keyword{flex}
\keyword{page}
