% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateCSSPropertyStyles}
\alias{generateCSSPropertyStyles}
\title{Creates styles for a css property.}
\usage{
generateCSSPropertyStyles(value, property, id, system)
}
\arguments{
\item{value}{Normalized attribute value list.}

\item{property}{The css attribute to generate rules for.}

\item{id}{The id of the parent element.}

\item{system}{A breakpoint system object to use.}
}
\value{
A vector of valid css strings.
}
\description{
Creates the required css statements for a specific css property. This
includes the rules for all breakpoints.
}
\keyword{generators}
\keyword{internal}
