% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateFlexChildrenCSS}
\alias{generateFlexChildrenCSS}
\title{Creates styles for the children of a css flex element.}
\usage{
generateFlexChildrenCSS(attributes, id, number_children, system)
}
\arguments{
\item{attributes}{Normalized attribute list.}

\item{id}{The id of the parent wrapper element.}

\item{number_children}{Number of child elements in the wrapper. Required to
generate rules for flex elements.}

\item{system}{A breakpoint system object to use.}
}
\value{
A vector of valid css strings.
}
\description{
Creates css required to position each child
element into the flex element.
}
\keyword{generators}
\keyword{internal}
