% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateFlexCSS}
\alias{generateFlexCSS}
\title{Creates styles for a css flex element.}
\usage{
generateFlexCSS(attributes, id, number_children, system)
}
\arguments{
\item{attributes}{Normalized attribute list.}

\item{id}{The id of the parent element.}

\item{number_children}{Number of child elements.}

\item{system}{A breakpoint system object to use.}
}
\value{
A vector of valid css strings.
}
\description{
Creates the requires css statements for a specific set of flex
attributes. This includes the css required to position each child
element into the flex element.
}
\keyword{generators}
\keyword{internal}
