% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateGridAreaCSS}
\alias{generateGridAreaCSS}
\title{Creates styles for css grid areas.}
\usage{
generateGridAreaCSS(areas, id)
}
\arguments{
\item{areas}{Vector of strings with the area names.}

\item{id}{The id of the parent wrapper element.}
}
\value{
A vector of valid css strings.
}
\description{
Creates the requires css statements for a specific set of grid areas.
This includes the css required to position each child element into
each of the named grid area.
}
\keyword{generators}
\keyword{internal}
