% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateGridCSS}
\alias{generateGridCSS}
\title{Generates all required css for a set of attributes, for a grid wrapper.}
\usage{
generateGridCSS(attributes, id, unique_areas, breakpoint_system)
}
\arguments{
\item{attributes}{Normalized attribute list to generate css from.}

\item{id}{The id of the parent wrapper element.}

\item{unique_areas}{Unique names of areas that the grid contains.}

\item{breakpoint_system}{Media breakpoints to use.}
}
\value{
A string with all placeholders replaced.
}
\description{
Generates all required css for a set of attributes, for a grid wrapper.
}
