% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateID}
\alias{generateID}
\title{Generates a HTML valid ID.
HTML IDs should follow specific standards:
  - The ID must start with a letter (a-z or A-Z).
  - All subsequent characters can be letters, numbers (0-9), hyphens (-),
    underscores (_), colons (:), and periods (.).
  - Each ID must be unique within the document.}
\usage{
generateID()
}
\value{
A valid CSS id.
}
\description{
Generates a HTML valid ID.
HTML IDs should follow specific standards:
  - The ID must start with a letter (a-z or A-Z).
  - All subsequent characters can be letters, numbers (0-9), hyphens (-),
    underscores (_), colons (:), and periods (.).
  - Each ID must be unique within the document.
}
