% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{getBreakpointSystem}
\alias{getBreakpointSystem}
\title{Returns a object form of a registered breakpoint system by its name or,
the currently active breakpoint system if no system name is provided.}
\usage{
getBreakpointSystem(system = NULL)
}
\arguments{
\item{system}{The name of a registered breakpoint system, or NULL if looking
for the currently active breakpoint system}
}
\value{
A breakpoint system object.
}
\description{
Returns a object form of a registered breakpoint system by its name or,
the currently active breakpoint system if no system name is provided.
}
\keyword{breakpoints}
