% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{getTemplate}
\alias{getTemplate}
\title{Get a registered template}
\usage{
getTemplate(name, type)
}
\arguments{
\item{name}{The name of a registered template.}

\item{type}{The type of css grid for which the template can be used.}
}
\value{
A template object.
}
\description{
Returns a object form of a registered template by its name and type.
}
\keyword{templates}
