% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{gridTemplate}
\alias{gridTemplate}
\title{Create a grid template}
\usage{
gridTemplate(
  name,
  type = c("grid", "flex"),
  ...,
  breakpoint_system = getBreakpointSystem(),
  description = NULL
)
}
\arguments{
\item{name}{A string with the name that identifies the template.}

\item{type}{The type of css grid for which the template can be used.
Value must be either "grid" or "flex".}

\item{...}{Collection of valid arguments that can be passed to a panel of
the given type (see \link{gridPanel} and \link{flexPanel} for all options)}

\item{breakpoint_system}{Breakpoint system to use.}

\item{description}{Optional description with information.
about the template. Can be used to pass on any additional relevant
information (For example its origin or connected frameworks).}
}
\value{
A template object.
}
\description{
Creates a imola template as an object for future use. Depending on the
given type, the template will then be available to be used as an
argument to a panel or page function of that specific type. Templates are
collections of arguments that can be grouped and stored for later usage
via the "template" argument of panel and page functions.
}
\keyword{templates}
