% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{importBreakpointSystem}
\alias{importBreakpointSystem}
\title{Imports a breakpoint system from a file.
Breakpoint systems can be exported into a file format by using
exportBreakpointSystem()}
\usage{
importBreakpointSystem(path)
}
\arguments{
\item{path}{A file path of the file to import, including the file
name and extension. The file name must end with a ".yaml" extension.}
}
\value{
A breakpoint system object.
}
\description{
Imports a breakpoint system from a file.
Breakpoint systems can be exported into a file format by using
exportBreakpointSystem()
}
\keyword{breakpoints}
