% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{importTemplate}
\alias{importTemplate}
\title{Import a template}
\usage{
importTemplate(path)
}
\arguments{
\item{path}{The file path of the file to import, including the file
name and extension. The file name must end with a \code{.yaml} extension.}
}
\value{
A template object.
}
\description{
Imports a template from a file.
Templates can be exported into a file format by using
\link{exportTemplate}
}
\keyword{templates}
