% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is.breakpoint}
\alias{is.breakpoint}
\title{Tests whether the object is an instance of a imola breakpoint.}
\usage{
is.breakpoint(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
TRUE or FALSE depending if the object is a imola breakpoint.
}
\description{
Tests whether the object is an instance of a imola breakpoint.
}
\keyword{internal}
\keyword{validators}
