% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is.breakpointSystem}
\alias{is.breakpointSystem}
\title{Tests whether the object is an instance of a imola breakpoint system.}
\usage{
is.breakpointSystem(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
TRUE or FALSE depending if the object is a imola breakpoint system.
}
\description{
Tests whether the object is an instance of a imola breakpoint system.
}
\keyword{internal}
\keyword{validators}
