% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is.template}
\alias{is.template}
\title{Tests whether the object is an instance of a imola template.}
\usage{
is.template(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
TRUE or FALSE depending if the object is a imola template.
}
\description{
Tests whether the object is an instance of a imola template.
}
\keyword{internal}
\keyword{validators}
