% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSystems.R
\name{listBreakpointSystems}
\alias{listBreakpointSystems}
\title{List registered breakpoint systems}
\usage{
listBreakpointSystems()
}
\value{
A named list of css templates and specific values.
}
\description{
Lists all available breakpoint systems.
}
\keyword{breakpoint_system}
\keyword{breakpoints}
