% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{listTemplates}
\alias{listTemplates}
\title{List registered templates}
\usage{
listTemplates(type = NULL)
}
\arguments{
\item{type}{Optional argument for what type of css templates to return.
value must be either "grid" or "flex". If no type is given, all templates
of all types are returned.}
}
\value{
A named list of css templates and specific values.
}
\description{
Lists all available grid and flex templates. If type is given, returns only
templates for the given grid type.
}
\keyword{templates}
