% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeAttribute}
\alias{normalizeAttribute}
\title{Normalize a attribute}
\usage{
normalizeAttribute(attribute, simplify = FALSE)
}
\arguments{
\item{attribute}{A attribute value.}

\item{simplify}{Should each attribute value be simplified into a single
string.}
}
\value{
A named list.
}
\description{
Converts the values of an attribute passed to a grid or flex function into
a normalized named list.
Does nothing if the attribute is already a named list in the correct format.
}
\keyword{internal}
\keyword{normalizer}
