% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeAttributes}
\alias{normalizeAttributes}
\title{Applies the normalizeAttribute() function to a full attribute list.}
\usage{
normalizeAttributes(attributes)
}
\arguments{
\item{attributes}{The values to process}
}
\value{
A named list.
}
\description{
Applies the normalizeAttribute() function to a full attribute list.
}
