% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.imola.breakpoint.system}
\alias{print.imola.breakpoint.system}
\title{Custom print function for a breakpoint system object.}
\usage{
\method{print}{imola.breakpoint.system}(x, ...)
}
\arguments{
\item{x}{the object to print.}

\item{...}{Additional arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Custom print function for a breakpoint system object.
}
\keyword{printer}
