% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.imola.template}
\alias{print.imola.template}
\title{Custom print function for a template object.}
\usage{
\method{print}{imola.template}(x, ...)
}
\arguments{
\item{x}{the object to print.}

\item{...}{Additional arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Custom print function for a template object.
}
\keyword{printer}
