% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readSettingsFile}
\alias{readSettingsFile}
\title{Reads the content of a YAML settings file from the package directory.}
\usage{
readSettingsFile(file)
}
\arguments{
\item{file}{The file name to read. Settings files are stored in the package
installation directory and include diferent settings and options.}
}
\value{
A list object containing the content of the settings YAML file
}
\description{
Reads the content of a YAML settings file from the package directory.
}
