% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{registerBreakpoint}
\alias{registerBreakpoint}
\title{Adds a new breakpoint entry to the currelty active media breakpoints.}
\usage{
registerBreakpoint(name, min = NULL, max = NULL)
}
\arguments{
\item{name}{The name of the entry to remove}

\item{min}{The minimum screen width (in pixels) when the rule is active}

\item{max}{The maximum screen width (in pixels) when the rule is active}
}
\value{
No return value, called for side effects
}
\description{
Adds a new breakpoint entry to the currelty active media breakpoints.
}
