% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{registerBreakpointSystem}
\alias{registerBreakpointSystem}
\title{Registers a breakpoint system object to make it available globally in
getOption("imola.breakpoints"). After registered it can be retrieved
anywhere using getBreakpointSystem().}
\usage{
registerBreakpointSystem(system)
}
\arguments{
\item{system}{A breakpoint system object generated with breakpointSystem().}
}
\value{
No return value, called for side effects.
}
\description{
Registers a breakpoint system object to make it available globally in
getOption("imola.breakpoints"). After registered it can be retrieved
anywhere using getBreakpointSystem().
}
\keyword{breakpoints}
