% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{registerTemplate}
\alias{registerTemplate}
\title{Register a template}
\usage{
registerTemplate(template)
}
\arguments{
\item{template}{A template object generated with \link{gridTemplate}.}
}
\value{
No return value, called for side effects.
}
\description{
Registers a template object to make it available globally in
getOption("imola.templates"). After registered it can be retrieved
anywhere using \link{getTemplate}.
}
\keyword{templates}
