% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{removeBreakpoint}
\alias{removeBreakpoint}
\title{Removes a breakpoint from a breakpoint system object by name.}
\usage{
removeBreakpoint(system, breakpoint)
}
\arguments{
\item{system}{A breakpoint system object generated with breakpointSystem().}

\item{breakpoint}{The name of a breakpoint in the given system.}
}
\value{
A breakpoint system object.
}
\description{
Removes a breakpoint from a breakpoint system object by name.
}
\keyword{breakpoints}
