% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{setActiveBreakpointSystem}
\alias{setActiveBreakpointSystem}
\title{Sets the current globally active breakpoint system.}
\usage{
setActiveBreakpointSystem(system)
}
\arguments{
\item{system}{The name of a registered breakpoint system or a breakpoint
system object generated with breakpointSystem().}
}
\value{
A breakpoint system object.
}
\description{
Sets the current globally active breakpoint system.
}
\keyword{breakpoints}
