% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{setBreakpointSystem}
\alias{setBreakpointSystem}
\title{Sets the current active media breakpoints. By default the default_system
  setting values are used, but the list can be customized by using the
  registerBreakpoint() and unregisterBreakpoint() functions.}
\usage{
setBreakpointSystem(system)
}
\arguments{
\item{system}{The name of the media breakpoints system to use. Existing
systems can be found under getOption("imola.breakpoints")}
}
\value{
A named list of media breakpoints options.
}
\description{
Sets the current active media breakpoints. By default the default_system
  setting values are used, but the list can be customized by using the
  registerBreakpoint() and unregisterBreakpoint() functions.
}
