% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{unregisterBreakpoint}
\alias{unregisterBreakpoint}
\title{Allows removing an entry from the current activeBreakpoints.}
\usage{
unregisterBreakpoint(name)
}
\arguments{
\item{name}{The name of the entry to remove}
}
\value{
No return value, called for side effects
}
\description{
Allows removing an entry from the current activeBreakpoints.
}
