% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{unregisterBreakpointSystem}
\alias{unregisterBreakpointSystem}
\title{Unregisters a globally registered breakpoint system in
getOption("imola.breakpoints").}
\usage{
unregisterBreakpointSystem(system)
}
\arguments{
\item{system}{The name of a registered breakpoint system, available in
getOption("imola.breakpoints").}
}
\value{
No return value, called for side effects.
}
\description{
Unregisters a globally registered breakpoint system in
getOption("imola.breakpoints").
}
\keyword{breakpoints}
