% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{unregisterTemplate}
\alias{unregisterTemplate}
\title{Unregister a template}
\usage{
unregisterTemplate(name, type)
}
\arguments{
\item{name}{A string with the name of a registered template.
Registered templates are available in \code{getOption("imola.templates")}.}

\item{type}{The type of css grid for which the template can be used.}
}
\value{
No return value, called for side effects.
}
\description{
Removes a globally registered template from
\code{getOption("imola.templates")}.
}
\keyword{templates}
