% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valueToCSS}
\alias{valueToCSS}
\title{Converts a R List or vector object into a valid css string. Used primarily
  to convert normalized attribute values into css values during processing.}
\usage{
valueToCSS(value, property)
}
\arguments{
\item{value}{The list or vector with the values to be converted into css}

\item{property}{The target css property for which the value will be used.}
}
\value{
string containing a valid css value.
}
\description{
Converts a R List or vector object into a valid css string. Used primarily
  to convert normalized attribute values into css values during processing.
}
