% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{addBreakpoint}
\alias{addBreakpoint}
\title{Add a breakpoint to a breakpoint system}
\usage{
addBreakpoint(system, breakpoint)
}
\arguments{
\item{system}{A breakpoint system object created with \link{breakpointSystem}.}

\item{breakpoint}{A breakpoint created with \link{breakpoint}.}
}
\value{
A breakpoint system object.
}
\description{
Adds a breakpoint to a breakpoint system object.
}
\keyword{breakpoint}
\keyword{breakpoints}
