% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSystems.R
\name{breakpointSystem}
\alias{breakpointSystem}
\title{Create a breakpoint system}
\usage{
breakpointSystem(name, ..., description = NULL)
}
\arguments{
\item{name}{A string with the name that identifies the breakpoint system.}

\item{...}{One or more breakpoint objects created with \link{breakpoint}.}

\item{description}{Optional description with information.
about the breakpoint system. Can be used to pass on any information
regarding the system (For example its origin or connected frameworks).}
}
\value{
A breakpoint system object.
}
\description{
Creates a breakpoint system object containing all
the information about the system, including its name and set of
available breakpoints.
}
\keyword{breakpoint_system}
\keyword{breakpoints}
