% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSystems.R
\name{exportBreakpointSystem}
\alias{exportBreakpointSystem}
\title{Export a breakpoint system}
\usage{
exportBreakpointSystem(system, path)
}
\arguments{
\item{system}{A string with the name of a registered breakpoint system, or a
breakpoint system object generated with \link{breakpointSystem}.}

\item{path}{The file path where to export the system to, including the file
name and extension. The file name must end with a \code{.yaml} extension.}
}
\value{
No return value, called for side effects.
}
\description{
Exports a breakpoint system into a file for storage and later usage.
Exported systems can be retrieved from their file form by using
\link{importBreakpointSystem}.
}
\keyword{breakpoint_system}
\keyword{breakpoints}
