% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateCSSPropertyStyles}
\alias{generateCSSPropertyStyles}
\title{Generates the requires css statements for a specific css property. It will
  iterated over all breakpoints in the given value and return all statements
  for all breakpoints in a vector format.}
\usage{
generateCSSPropertyStyles(value, property, id, breakpoint_system)
}
\arguments{
\item{value}{Normalized attribute value list to generate css from.}

\item{property}{The type of grid to generate css for.}

\item{id}{The id of the parent wrapper element.}

\item{breakpoint_system}{Media breakpoints to use.}
}
\value{
A vector of valid css strings.
}
\description{
Generates the requires css statements for a specific css property. It will
  iterated over all breakpoints in the given value and return all statements
  for all breakpoints in a vector format.
}
