% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateFlexCSS}
\alias{generateFlexCSS}
\title{Generates all required css for a set of attributes, for a flex wrapper.}
\usage{
generateFlexCSS(attributes, id, number_children, breakpoint_system)
}
\arguments{
\item{attributes}{Normalized attribute list to generate css from.}

\item{id}{The id of the parent wrapper element.}

\item{number_children}{Number of child elements in the wrapper. Required to
generate rules for flex elements.}

\item{breakpoint_system}{Media breakpoints to use.}
}
\value{
A string with all placeholders replaced.
}
\description{
Generates all required css for a set of attributes, for a flex wrapper.
}
