% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateGridAreaCSS}
\alias{generateGridAreaCSS}
\title{Generates the requires css statements for a specific set of grid areas. This
  includes the css required to position each child element into each of the
  named grid area.}
\usage{
generateGridAreaCSS(areas, id)
}
\arguments{
\item{areas}{Unique names of areas for which css rules must be generated.}

\item{id}{The id of the parent wrapper element.}
}
\value{
A vector of valid css strings.
}
\description{
Generates the requires css statements for a specific set of grid areas. This
  includes the css required to position each child element into each of the
  named grid area.
}
