% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generateGridCSS}
\alias{generateGridCSS}
\title{Creates styles for css grid attributes.}
\usage{
generateGridCSS(attributes, id, areas, system)
}
\arguments{
\item{attributes}{Normalized attribute list.}

\item{id}{The id of the parent wrapper element.}

\item{areas}{Area names in the grid.}

\item{system}{A breakpoint system object to use.}
}
\value{
A vector of valid css strings.
}
\description{
Creates styles for css grid attributes.
}
\keyword{generators}
\keyword{internal}
