% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSystems.R
\name{getBreakpointSystem}
\alias{getBreakpointSystem}
\title{Get a registered breakpoint system}
\usage{
getBreakpointSystem(name = NULL)
}
\arguments{
\item{name}{A string with the name of a registered breakpoint system, or
\code{NULL} if looking for the currently active breakpoint system.}
}
\value{
A breakpoint system object.
}
\description{
Returns a breakpoint system object of a registered breakpoint system by
its name or, the currently active breakpoint system if no system name is
provided.
}
\keyword{breakpoint_system}
\keyword{breakpoints}
