% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSystems.R
\name{importBreakpointSystem}
\alias{importBreakpointSystem}
\title{Import a breakpoint system}
\usage{
importBreakpointSystem(path)
}
\arguments{
\item{path}{The file path of the file to import, including the file
name and extension. The file name must end with a \code{.yaml} extension.}
}
\value{
A breakpoint system object.
}
\description{
Imports a breakpoint system from a file.
Breakpoint systems can be exported into a file format using
\link{exportBreakpointSystem}.
}
\keyword{breakpoint_system}
\keyword{breakpoints}
