% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeAttribute}
\alias{normalizeAttribute}
\title{Converts non named list attributes into a named list.
Does nothing if the attribute is already a list in the correct format.}
\usage{
normalizeAttribute(attribute, simplify = TRUE)
}
\arguments{
\item{attribute}{The value to process}

\item{simplify}{Boolean flag if the attribute should be simplified into
single strings.}
}
\value{
A named list.
}
\description{
Converts non named list attributes into a named list.
Does nothing if the attribute is already a list in the correct format.
}
