% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeAttributes}
\alias{normalizeAttributes}
\title{Normalize a list of attributes}
\usage{
normalizeAttributes(attributes)
}
\arguments{
\item{attributes}{A named list of attribute values.}
}
\value{
A named list.
}
\description{
Iterates through the list of attributes and applies \link{normalizeAttribute} to
each of them.
}
\keyword{internal}
\keyword{normalizer}
