% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSystems.R
\name{registerBreakpointSystem}
\alias{registerBreakpointSystem}
\title{Register a breakpoint system}
\usage{
registerBreakpointSystem(system)
}
\arguments{
\item{system}{A breakpoint system object created with \link{breakpointSystem}.}
}
\value{
No return value, called for side effects.
}
\description{
Registers a breakpoint system object to make it available globally in
\code{getOption("imola.breakpoint.systems")}. After registered it can be retrieved
anywhere using \link{getBreakpointSystem}.
}
\keyword{breakpoint_system}
\keyword{breakpoints}
