% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{removeBreakpoint}
\alias{removeBreakpoint}
\title{Remove a breakpoint from a breakpoint system}
\usage{
removeBreakpoint(system, name)
}
\arguments{
\item{system}{A breakpoint system object created with \link{breakpointSystem}.}

\item{name}{A string with the name of a breakpoint in the given system.}
}
\value{
A breakpoint system object.
}
\description{
Removes a breakpoint from a breakpoint system object by name.
}
\keyword{breakpoint}
\keyword{breakpoints}
