% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSystems.R
\name{setActiveBreakpointSystem}
\alias{setActiveBreakpointSystem}
\title{Set the active breakpoint system}
\usage{
setActiveBreakpointSystem(system)
}
\arguments{
\item{system}{A string with the name of a registered breakpoint system, or a
breakpoint system object generated with \link{breakpointSystem}. If a breakpoint
system object is used, it will be registered as well.}
}
\value{
A breakpoint system object.
}
\description{
Sets the current globally active breakpoint system. The active breakpoint
system is used for grid function as the default system if no system is
provided as an argument.
}
\keyword{breakpoint_system}
\keyword{breakpoints}
