% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{unregisterTemplate}
\alias{unregisterTemplate}
\title{Deletes an existing css template from the available list of templates for the
  given grid type. Templates are collections of arguments
  that can be grouped and stored for later usage via the "template"
  argument of panel and page functions.}
\usage{
unregisterTemplate(type, name)
}
\arguments{
\item{type}{The type of css grid of the template to remove.}

\item{name}{The name of the tempalte to remove.}
}
\value{
No return value, called for side effects
}
\description{
Deletes an existing css template from the available list of templates for the
  given grid type. Templates are collections of arguments
  that can be grouped and stored for later usage via the "template"
  argument of panel and page functions.
}
